Update from original

+++ images/base/scripts/common/setup.sh
@@ -18,23 +18,37 @@ apt-get install --yes git
 
 # Set Go environment variables needed by other scripts
 export GOPATH="/opt/gopath"
-export GOROOT="/opt/go"
-PATH=$GOROOT/bin:$GOPATH/bin:$PATH
 
 #install golang
 #apt-get install --yes golang
 mkdir -p $GOPATH
+MACHINE=`uname -m`
+if [ x$MACHINE = xs390x ]
+then
+   apt-get install --yes golang
+   export GOROOT="/usr/lib/go-1.6"
+elif [ x$MACHINE = xppc64 ]
+then
+   echo "TODO: Add PPC support"
+   exit
+elif [ x$MACHINE = xx86_64 ]
+then
+   export GOROOT="/opt/go"
 
-#ARCH=`uname -m | sed 's|i686|386|' | sed 's|x86_64|amd64|'`
-ARCH=amd64
-GO_VER=1.6
+   #ARCH=`uname -m | sed 's|i686|386|' | sed 's|x86_64|amd64|'`
+   ARCH=amd64
+   GO_VER=1.6
 
-cd /tmp
-wget --quiet --no-check-certificate https://storage.googleapis.com/golang/go$GO_VER.linux-${ARCH}.tar.gz
-tar -xvf go$GO_VER.linux-${ARCH}.tar.gz
-mv go $GOROOT
-chmod 775 $GOROOT
-rm go$GO_VER.linux-${ARCH}.tar.gz
+   cd /tmp
+   wget --quiet --no-check-certificate https://storage.googleapis.com/golang/go$GO_VER.linux-${ARCH}.tar.gz
+   tar -xvf go$GO_VER.linux-${ARCH}.tar.gz
+   mv go $GOROOT
+   chmod 775 $GOROOT
+   rm go$GO_VER.linux-${ARCH}.tar.gz
+fi
+
+PATH=$GOROOT/bin:$GOPATH/bin:$PATH
+
 cat <<EOF >/etc/profile.d/goroot.sh
 export GOROOT=$GOROOT
 export GOPATH=$GOPATH
@@ -43,25 +57,22 @@ EOF
 
 
 # Install NodeJS
-#cd /tmp
-#ver=0.12.7 #Replace this with the latest version available
-#wget -c http://nodejs.org/dist/v$ver/node-v$ver.tar.gz #This is to download the source code.
-#tar -xzf node-v$ver.tar.gz
-#rm node-v$ver.tar.gz
-#cd node-v$ver
-#./configure && make && sudo make install
-#cd /tmp && rm -rf node-v$ver
 
-NODE_VER=0.12.7
-NODE_PACKAGE=node-v$NODE_VER-linux-x64.tar.gz
-TEMP_DIR=/tmp
-SRC_PATH=$TEMP_DIR/$NODE_PACKAGE
+if [ x$MACHINE = xs390x ]
+then
+    apt-get install --yes nodejs
+else
+    NODE_VER=0.12.7
+    NODE_PACKAGE=node-v$NODE_VER-linux-x64.tar.gz
+    TEMP_DIR=/tmp
+    SRC_PATH=$TEMP_DIR/$NODE_PACKAGE
 
-cd $TEMP_DIR
-# First remove any prior packages downloaded in case of failure
-rm -f node*.tar.gz
-wget --quiet https://nodejs.org/dist/v$NODE_VER/$NODE_PACKAGE
-cd /usr/local && sudo tar --strip-components 1 -xzf $SRC_PATH
+    # First remove any prior packages downloaded in case of failure
+    cd $TEMP_DIR
+    rm -f node*.tar.gz
+    wget --quiet https://nodejs.org/dist/v$NODE_VER/$NODE_PACKAGE
+    cd /usr/local && sudo tar --strip-components 1 -xzf $SRC_PATH
+fi
 
 # Install GRPC
 
@@ -74,7 +85,7 @@ cd /usr/local && sudo tar --strip-components 1 -xzf $SRC_PATH
 cd /tmp
 git clone https://github.com/google/protobuf.git
 cd protobuf
-git checkout 12fb61b292d7ec4cb14b0d60e58ed5c35adda3b7
+git checkout v3.0.0-beta-3
 #unzip needed for ./autogen.sh
 apt-get install -y unzip
 apt-get install -y autoconf
@@ -90,9 +101,14 @@ apt-get install -y build-essential libtool
 #./configure
 ./configure --prefix=/usr
 
-make
-make check
-make install
+if [ x$MACHINE = xs390x ]
+then
+    echo FIXME: protobufs wont compile on 390, missing atomic call
+else
+    make
+    make check
+    make install
+fi
 export LD_LIBRARY_PATH=/usr/local/lib:$LD_LIBRARY_PATH
 cd ~/
 
@@ -102,6 +118,13 @@ cd /tmp
 git clone https://github.com/facebook/rocksdb.git
 cd rocksdb
 git checkout tags/v4.1
+if [ x$MACHINE = xs390x ]
+then
+    echo There were some bugs in 4.1 for x/p, dev stream has the fix, living dangereously, fixing in place
+    sed -i -e "s/-march=native/-march=z196/" build_tools/build_detect_platform
+    sed -i -e "s/-momit-leaf-frame-pointer/-DDUMBDUMMY/" Makefile
+fi
+
 PORTABLE=1 make shared_lib
 INSTALL_PATH=/usr/local make install-shared
 ldconfig
